Add builder for benches
Also adds a "sane" case for block processing
